backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
pyjwt==2.8.0
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0

# Validation & Sanitization
pydantic==2.10.5
pydantic-settings==2.7.0
email-validator==2.2.0
bleach==6.1.0

# Rate Limiting
slowapi==0.1.9

# File handling
python-magic==0.4.27



# Environment variables
python-dotenv==1.0.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests==2.31.0
colorama==0.4.6
